Genauigkeit angepasst

+++ File2.java
@@ -1,7 +1,7 @@
 public class File2{
 	public static void main(String[] args){
 		System.out.println("Hallo, ich bin File2!");
-		int n = 10;
-		int x = n + 1;
+		double n = 10.0;
+		int x = (int)(n + 1);
 	}
 }
